changed column type

+++ estimate_delivery_time/database/migrations/2021_04_05_150610_create_zip_codes_table.php
@@ -15,7 +15,7 @@ class CreateZipCodesTable extends Migration
     {
         Schema::create('zip_codes', function (Blueprint $table) {
             $table->id();
-            $table->integer("zip_code");
+            $table->bigInteger("zip_code");
         });
     }
